+++ applications/fishing-map/src/features/dataviews/dataviews.utils.ts
@@ -11,6 +11,8 @@ import {
   DEFAULT_VESSEL_DATAVIEW_ID,
 } from 'data/workspaces'
 
+// used in workspaces with encounter events layers
+export const ENCOUNTER_EVENTS_SOURCE_ID = 'encounter-events'
 export const DATAVIEW_INSTANCE_PREFIX = 'vessel-'
 export const ENVIRONMENTAL_LAYER_PREFIX = 'environment-'
 export const CONTEXT_LAYER_PREFIX = 'context-'
@@ -39,6 +41,7 @@ export const getVesselDataviewInstance = (
     id: `${DATAVIEW_INSTANCE_PREFIX}${vessel.id}`,
     dataviewId: DEFAULT_VESSEL_DATAVIEW_ID,
     config: {
+      // TODO pick a not used color
       color: TrackColorBarOptions[Math.floor(Math.random() * TrackColorBarOptions.length)].value,
     },
     datasetsConfig,