+++ applications/fishing-map/src/index.tsx
@@ -4,7 +4,6 @@ import { Provider } from 'react-redux'
 import { RecoilRoot } from 'recoil'
 import App from 'features/app/App'
 import store from './store'
-import * as serviceWorker from './serviceWorker'
 
 import './features/i18n/i18n'
 import '@globalfishingwatch/ui-components/dist/base.css'
@@ -27,8 +26,3 @@ render(
   </React.StrictMode>,
   document.getElementById('root')
 )
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister()